supply-chain-risk-index/requirements.txt
jupyter
matplotlib
networkx
numpy
pandas
python-dotenv
requests
scikit-learn
seaborn
spacy
textblob
